+++ test/function/expression/eval.test.js
@@ -1,11 +1,12 @@
 // test eval
-var assert = require('assert'),
-    approx = require('../../../tools/approx'),
-    error = require('../../../lib/error/index'),
-    math = require('../../../index')(),
-    Complex = math.type.Complex,
-    Matrix = math.type.Matrix,
-    Unit = math.type.Unit;
+var assert = require('assert');
+var approx = require('../../../tools/approx');
+var error = require('../../../lib/error/index');
+var math = require('../../../index')();
+var Complex = math.type.Complex;
+var Matrix = math.type.Matrix;
+var Unit = math.type.Unit;
+var ResultSet = math.type.ResultSet;
 
 describe('eval', function() {
 
@@ -22,9 +23,9 @@ describe('eval', function() {
   });
 
   it('should eval a series of expressions', function() {
-    assert.deepEqual(math.eval('a=3\nb=4\na*b'), [3, 4, 12]);
-    assert.deepEqual(math.eval('f(x) = a * x; a=2; f(4)'), [8]);
-    assert.deepEqual(math.eval('b = 43; b * 4'), [172]);
+    assert.deepEqual(math.eval('a=3\nb=4\na*b'), new ResultSet([3, 4, 12]));
+    assert.deepEqual(math.eval('f(x) = a * x; a=2; f(4)'), new ResultSet([8]));
+    assert.deepEqual(math.eval('b = 43; b * 4'), new ResultSet([172]));
   });
 
   it('should throw an error if wrong number of arguments', function() {